Added AI Recommendations Feature
Added AI recommendations feature. Please retrieve a new Google Gemini API key and paste it into the right location in the aiRecommendations.tsx file for the feature to work in the simulator.

Some bugs I noticed: The Add Exercises and Get AI Recommendations button still appear even on the page where sets and weights can be inputted (it is not needed there). Delete button for the routines still does not work as well.